Update status.php
The biggest change is the addition of the getExtensionName method, which uses the "sip show user 1234" command to get the name of the person at the chosen extension.  Additionally, I made some minor cosmetic changes, such as moving the Channel information into the row's Title attribute (so it appears when you hover), and formatting the phone number into a U.S.-style number.